Testing Realm and Room (Instrusment Testing)

+++ app/src/main/java/com/fpremake/screens_post_login/screen_dashboard/data/room/Parent.kt
@@ -8,6 +8,6 @@ import androidx.room.PrimaryKey
 data class Parent(
     @PrimaryKey(autoGenerate = true)
     @ColumnInfo(name = "parentId")
-    var parentId:Long = 0,
+    var parentId:Int = 0,
     @ColumnInfo(name = "name") var name: String? = "",
 )
